Use **row** in the organizing and executing simulations tutorial (#160)
* First draft of row tutorial.

* Remove unused project file.

* ruff format

* Fix link.

* Test with python 3.13

* pre-commit

* Fix cache key.

* Grammar pass.

* pre-commit

* Grammar and consistency.

* More tweaks.
