+++ .github/workflows/environments/update-lockfiles.sh
@@ -4,7 +4,7 @@
 
 rm *-conda-lock.yml
 
-for python_version in 3.12
+for python_version in 3.13
 do
   sed "s/python==.*/python=$python_version/g" environment.yaml > tmp.yaml || exit 2
   conda lock -f tmp.yaml -p linux-64 --lockfile py${python_version//.}-conda-lock.yml || exit 2
