+++ .github/workflows/test.yaml
@@ -41,7 +41,7 @@ jobs:
       uses: actions/cache/restore@5a3ec84eff668545956fd18022155c47e93e2684 # v4.2.3
       with:
         path: install
-        key: hoomd-blue-${{ steps.hoomd.outputs.sha }}
+        key: hoomd-blue-${{ steps.hoomd.outputs.sha }}-py313
 
     - name: Checkout notebooks
       uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683 # v4.2.2
@@ -52,7 +52,7 @@ jobs:
       uses: mamba-org/setup-micromamba@0dea6379afdaffa5d528b3d1dabc45da37f443fc # v2.0.4
       with:
         environment-name: test
-        environment-file: notebooks/.github/workflows/environments/py312-conda-lock.yml
+        environment-file: notebooks/.github/workflows/environments/py313-conda-lock.yml
         micromamba-root-path: ${{ github.workspace }}/micromamba
         micromamba-version: '2.0.4-0'
